backend/requirements.txt
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
asyncpg>=0.29.0
pydantic-settings>=2.3.4
certifi>=2024.7.4
httpx>=0.27.0
tenacity>=8.3.0
sse-starlette>=2.0.0
pytest>=8.2.0

